refactor: rename redux-states/*/action-constants.ts to action-types.ts
