+++ src/redux-states/project/actions.ts
@@ -14,7 +14,7 @@ import {
   SHOW_LOADER,
   UPDATE_PROJECT,
   WRITE_PROJECT_FAILURE,
-} from './action-constants';
+} from './action-types';
 import { IProjectState } from './model';
 import SAMPLE from './localStorage';
