+++ src/redux-states/task-list/actions.ts
@@ -1,6 +1,6 @@
 import ITask from 'Entities/task';
 import * as StateTypes from 'States/types';
-import { DESELECT_TASK, HIDE_EDIT, SELECT_TASK, SHOW_EDIT } from './action-constants';
+import { DESELECT_TASK, HIDE_EDIT, SELECT_TASK, SHOW_EDIT } from './action-types';
 
 export const deselectTask = (): StateTypes.IAction<undefined> => ({
   type: DESELECT_TASK,
