+++ src/redux-states/task-list/index.ts
@@ -1,5 +1,5 @@
 import * as taskListActions from './actions';
-import * as taskListConstants from './action-constants';
+import * as taskListConstants from './action-types';
 import * as taskListTypes from './model';
 import taskListReducer from './reducer';
